# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Piece-Meal App
# during dev, should point to your local machine
# host: piecemeal-api.herokuapp.com
host: localhost:10010
...
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  # - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /clients/{user_id}/recipes:
    x-swagger-router-controller: recipes
    get:
      description: get a client's recipes
      operationId: getClientRecipes
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          description: user's id
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: add a new recipe to the client's list of recipes
      operationId: addClientRecipe
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          description: client's id
          required: true

        - name: request
          in: body
          required: true
          schema:
            properties:
              recipe_id:
                type: integer
            required:
              - recipe_id
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /recipes:
    x-swagger-router-controller: recipes
    get:
      description: get the recipes list
      operationId: getRecipesList
      # parameters:
      #   - name: token
      #     in: header
      #     description: Users authentication token
      #     required: true
      #     type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: add a new recipe
      operationId: postRecipe
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: recipe
          description: Recipe
          in: body
          required: true
          schema:
            $ref: "#/definitions/PostRecipe"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Recipe"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /random:
    x-swagger-router-controller: recipes
    get:
      description: get random recipes
      operationId: getRandomRecipes
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /recipes/{id}:
    x-swagger-router-controller: recipes
    get:
      description: get a recipe
      # define the type of response for Success "200" and Error
      operationId: getRecipeBookshelf
      # operationId: getRecipe
      parameters:
        # - name: token
        #   in: header
        #   description: Users authentication token
        #   required: true
        #   type: string

        - name: id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Recipe"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: update a recipe
      # define the parameters
      operationId: updateRecipe
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: id
          description: recipe id
          type: integer
          in: path
          required: true
        - name: recipe
          description: recipe
          in: body
          required: true
          schema:
            $ref: "#/definitions/PostRecipe"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Recipe"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"


    delete:
      description: delete a recipe
      # define the parameters
      operationId: deleteRecipe
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: id
          description: recipe id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Recipe"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clients/{user_id}/recipes/favorites:
    x-swagger-router-controller: recipes
    get:
      description: get client's favorite recipes
      operationId: getFavoriteRecipes
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /recipes/{recipe_id}/ratings:
    x-swagger-router-controller: recipes
    post:
      description: rate a recipe
      # define the type of response for Success "200" and Error
      operationId: rateRecipe
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: recipe_id
          type: integer
          in: path
          required: true

        - name: vote
          in: body
          schema:
            type: object
            required:
              - client_id
              - vote
            properties:
              client_id:
                type: integer
              vote:
                type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Recipe"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /search/recipes:
    x-swagger-router-controller: recipes
    get:
      description: get results from recipe search
      operationId: searchRecipes
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: text
          type: string
          in: query
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # /clients/{user_id}/recipes/{recipe_id}/rating:
  #   x-swagger-router-controller: clients
  #   get:
  #     description: to get client's rating for a recipe
  #     parameters:
  #       - name: user_id
  #         type: integer
  #         in: path
  #         required: true
  #       - name: recipe_id
  #         type: integer
  #         in: path
  #         required: true
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           $ref: "#/definitions/GetUserRecipeRatingResponse"
  #       default:
  #           description: Error
  #           schema:
  #             $ref: "#/definitions/ErrorResponse"
  #   post:
  #     description: to post/update client's rating
  #     parameters:
  #       - name: user_id
  #         type: integer
  #         in: path
  #         required: true
  #       - name: recipe_id
  #         in: path
  #         type: integer
  #         required: true
  #       - name: rating
  #         in: body
  #         schema:
  #           type: object
  #           required:
  #             - rating
  #           properties:
  #             rating:
  #               description: rating to post/update
  #               type: number
  #
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           $ref: "#/definitions/GeneralResponse"
  #       default:
  #           description: Error
  #           schema:
  #             $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Recipe:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      notes:
        type: string
      active:
        type: boolean
      ingredients:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            description:
              type: string
            active:
              type: boolean
            tags:
              type: array
              items:
                type: string
      instructions:
        type: array
        items:
          type: object
          properties:
            step_number:
              type: integer
            instructions:
              type: string
      image_url:
        type: string
        description: link to image of ingredient
  PostRecipe:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      instructions:
        type: array
        items:
          type: object
          properties:
            step_number:
              type: integer
            instructions:
              type: string
      ingredients:
        type: array
        items:
          type: integer
      tags:
        type: array
        items:
          type: string
      image_url:
        type: string
        description: link to image of ingredient
  GetRecipesListResponse:
    required:
      - recipes
    properties:
      recipes:
        type: array
        items:
          $ref: "#/definitions/Recipe"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: integer
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description

  # IngredientResponse:
  #   type: object
  #   properties:
  #     success:
  #       type: integer
  #       description: returns 1 if successful
  #     description:
  #       type: object
  #       properties:
  #       description: a short comment
  #   required:
  #     - success
  #     - description